qualify all integer sizes FixedTokenSupply and FixedNFTSet

+++ contracts/FixedNFTSet.sol
@@ -12,14 +12,14 @@ contract FixedNFTSet is ERC721, ERC721Enumerable, ERC165 {
 
 // The number of tokens is fixed during contract creation.
 // Zero/absent entries in tokenOwners take the defaultOwner value.
-uint private immutable tokenCountAndDefaultOwner;
+uint256 private immutable tokenCountAndDefaultOwner;
 
 // Each token (index/ID) has one owner at a time.
 // Zero/absent entries take the defaultOwner value.
-mapping(uint => address) private tokenOwners;
+mapping(uint256 => address) private tokenOwners;
 
 // Each token (index/ID) can be granted to a destination address.
-mapping(uint => address) private tokenApprovals;
+mapping(uint256 => address) private tokenApprovals;
 
 // The token-owner:token-operator:approval-flag entries are always true.
 mapping(address => mapping(address => bool)) private operatorApprovals;
@@ -27,7 +27,7 @@ mapping(address => mapping(address => bool)) private operatorApprovals;
 // Constructor mints n tokens, and transfers each token to the receiver address.
 // Token identifiers match their respective index, counting from 0 to n − 1.
 // Initial Transfer emission is omitted.
-constructor(uint n, address receiver) {
+constructor(uint256 n, address receiver) {
 	requireAddress(receiver);
 	tokenCountAndDefaultOwner = uint(uint160(receiver)) | (n << 160);
 }
@@ -38,7 +38,7 @@ function requireAddress(address a) internal pure {
 }
 
 // RequireToken denies any token index/ID that is not in this contract.
-function requireToken(uint indexOrID) internal view {
+function requireToken(uint256 indexOrID) internal view {
 	require(indexOrID < totalSupply(), "ERC-721 token \u2415");
 }
 
@@ -49,17 +49,17 @@ function supportsInterface(bytes4 interfaceID) public virtual override(ERC165) p
 	    || interfaceID == 0x01ffc9a7; // ERC165
 }
 
-function totalSupply() public override(ERC721Enumerable) view returns (uint) {
+function totalSupply() public override(ERC721Enumerable) view returns (uint256) {
 	return tokenCountAndDefaultOwner >> 160;
 }
 
 // Tokens are identified by their index one-to-one.
-function tokenByIndex(uint index) public override(ERC721Enumerable) view returns (uint) {
+function tokenByIndex(uint256 index) public override(ERC721Enumerable) view returns (uint256) {
 	requireToken(index);
 	return index;
 }
 
-function tokenOfOwnerByIndex(address owner, uint index) public override(ERC721Enumerable) view returns (uint tokenID) {
+function tokenOfOwnerByIndex(address owner, uint256 index) public override(ERC721Enumerable) view returns (uint256 tokenID) {
 	requireAddress(owner);
 	for (tokenID = 0; tokenID < totalSupply(); tokenID++) {
 		if (ownerOf(tokenID) == owner) {
@@ -73,11 +73,11 @@ function tokenOfOwnerByIndex(address owner, uint index) public override(ERC721En
 	revert("ERC-721 index exceeds balance");
 }
 
-function balanceOf(address owner) public override(ERC721) view returns (uint) {
+function balanceOf(address owner) public override(ERC721) view returns (uint256) {
 	requireAddress(owner);
-	uint balance = 0;
+	uint256 balance = 0;
 	// count owner matches
-	for (uint tokenID = 0; tokenID < totalSupply(); tokenID++) {
+	for (uint256 tokenID = 0; tokenID < totalSupply(); tokenID++) {
 		if (ownerOf(tokenID) == owner) {
 			++balance;
 		}
@@ -85,7 +85,7 @@ function balanceOf(address owner) public override(ERC721) view returns (uint) {
 	return balance;
 }
 
-function ownerOf(uint tokenID) public override(ERC721) view returns (address) {
+function ownerOf(uint256 tokenID) public override(ERC721) view returns (address) {
 	requireToken(tokenID);
 	address owner = tokenOwners[tokenID];
 	if (owner == address(0)) {
@@ -94,18 +94,18 @@ function ownerOf(uint tokenID) public override(ERC721) view returns (address) {
 	return owner;
 }
 
-function safeTransferFrom(address from, address to, uint tokenID, bytes calldata data) public override(ERC721) payable {
+function safeTransferFrom(address from, address to, uint256 tokenID, bytes calldata data) public override(ERC721) payable {
 	transferFrom(from, to, tokenID);
 	if (msg.sender == tx.origin) { // is contract
 		require(ERC721TokenReceiver(to).onERC721Received(msg.sender, from, tokenID, data) == ERC721TokenReceiver.onERC721Received.selector, "ERC721TokenReceiver mis");
 	}
 }
 
-function safeTransferFrom(address from, address to, uint tokenID) public override(ERC721) payable {
+function safeTransferFrom(address from, address to, uint256 tokenID) public override(ERC721) payable {
 	return this.safeTransferFrom(from, to, tokenID, "");
 }
 
-function transferFrom(address from, address to, uint tokenID) public override(ERC721) payable {
+function transferFrom(address from, address to, uint256 tokenID) public override(ERC721) payable {
 	address owner = ownerOf(tokenID); // checks token ID
 	require(from == owner, "ERC-721 from \u2415");
 	requireAddress(to);
@@ -124,7 +124,7 @@ function transferFrom(address from, address to, uint tokenID) public override(ER
 	emit Transfer(from, to, tokenID);
 }
 
-function approve(address to, uint tokenID) public override(ERC721) payable {
+function approve(address to, uint256 tokenID) public override(ERC721) payable {
 	address owner = ownerOf(tokenID); // checks token ID
 	require(msg.sender == owner || isApprovedForAll(owner, msg.sender), "ERC-721 sender deny");
 	tokenApprovals[tokenID] = to;
@@ -140,7 +140,7 @@ function setApprovalForAll(address operator, bool approved) public override(ERC7
 	emit ApprovalForAll(msg.sender, operator, approved);
 }
 
-function getApproved(uint tokenID) public override(ERC721) view returns (address operator) {
+function getApproved(uint256 tokenID) public override(ERC721) view returns (address operator) {
 	requireToken(tokenID);
 	return tokenApprovals[tokenID];
 }
